move update emission factor logic into db service
